requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.2.0
